Add files via upload
Made an Entity-Relationship (ER) diagram that is easy to read (RDC_ER_Diagram.pdf). Generated a database catalog (DB_Catalog.xlsx) that consists of all the entities, relationships, and attributes. Each attribute has a specified data type and constraint. If the attribute is constrained, it means that non-null data must be entered into that field. This will be useful for keeping track of how an application can interact with this database later on!