Add files via upload
Finished building the first version of the automated Excel to SQL database python program! This does not automate everything, there is still data that we will insert manually into our tables. These tables can be added to manually because they are not updated as frequently as orders are. This program DOES take the grunt work out of adding hundreds of rows of frequently updating data!